# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hyperlocal Community API",
    "description": "Hyperlocal is an open source community engagement platform\n\nSome useful links:\n- [Hyperlocal Nextjs repository](https://github.com/colpitts-dev/hyperlocal-nextjs)",
    "termsOfService": "http://hyperlocal.wiki/terms/",
    "contact": {
      "email": "apiteam@hyperlocal.wiki"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.1.0"
  },
  "externalDocs": {
    "description": "Find out more about Hyperlocal",
    "url": "http://hyperlocal.wiki"
  },
  "servers": [
    {
      "url": "https://dev.hyperlocal.wiki/api/v1"
    }
  ],
  "tags": [
    {
      "name": "person",
      "description": "A person belongs to many communities through their memberships"
    },
    {
      "name": "community",
      "description": "A Community is a group of people who share a common interest"
    },
    {
      "name": "membership",
      "description": "A membership belongs to a person and associates them with a community"
    }
  ],
  "paths": {
    "/people": {
      "post": {
        "tags": ["person"],
        "summary": "Create a new person",
        "description": "Create a new person in hyperlocal",
        "operationId": "createPerson",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "tags": ["person"],
        "summary": "Get all people",
        "description": "Get all people in hyperlocal",
        "operationId": "getPeople",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/people/{id}": {
      "get": {
        "tags": ["person"],
        "summary": "Get a person",
        "description": "Get a person by id",
        "operationId": "getPerson",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of person to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "patch": {
        "tags": ["person"],
        "summary": "Update a person",
        "description": "Update a person by id",
        "operationId": "updatePerson",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": ["person"],
        "summary": "Delete a person",
        "description": "Delete a person by id",
        "operationId": "deletePerson",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/communities": {
      "post": {
        "tags": ["community"],
        "summary": "Create a new community",
        "description": "Create a new community in hyperlocal",
        "operationId": "createCommunity",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "tags": ["community"],
        "summary": "Get all communities",
        "description": "Get all communities in hyperlocal",
        "operationId": "getCommunities",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/communities/{id}": {
      "get": {
        "tags": ["community"],
        "summary": "Get a community",
        "description": "Get a community by id",
        "operationId": "getCommunity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of community to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "patch": {
        "tags": ["community"],
        "summary": "Update a community",
        "description": "Update a community by id",
        "operationId": "updateCommunity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of community to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": ["community"],
        "summary": "Delete a community",
        "description": "Delete a community by id",
        "operationId": "deleteCommunity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of community to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/memberships": {
      "post": {
        "tags": ["membership"],
        "summary": "Create a new membership",
        "description": "Create a new membership in hyperlocal",
        "operationId": "createMembership",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "tags": ["membership"],
        "summary": "Get all memberships",
        "description": "Get all memberships in hyperlocal",
        "operationId": "getMemberships",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/memberships/{id}": {
      "get": {
        "tags": ["membership"],
        "summary": "Get a membership",
        "description": "Get a membership by id",
        "operationId": "getMembership",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of membership to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "patch": {
        "tags": ["membership"],
        "summary": "Update a membership",
        "description": "Update a membership by id",
        "operationId": "updateMembership",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of membership to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": ["membership"],
        "summary": "Delete a membership",
        "description": "Delete a membership by id",
        "operationId": "deleteMembership",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of membership to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    }
  }
}